Separate first responder state from first responder demand.
The former now communicates its changes using a simple callback system,
whereas the latter is still controlled using a binding.

This avoids some potential confusion with having a single enum where
only the text field itself can set the the "is or is not" first
responder state and only the "become or resign" state can be set
externally.