Fix bug with standard edit action handling where original might not get called.
The logic is supposed to be that if a custom override is provided
and it returns true, the original should be called, but if no override
is provided then the original should be called.

The previous implementation was calling the original using optional
chaining which meant the original would not be called if the override
was nil.

This change fixes that bug and also refactors out the duplicated
logic into a more general function.